refactor for more sanity
